gnrc_tcp: refactor tests
