+++ tests/gnrc_tcp/README.md
@@ -1,42 +1,7 @@
 Test description
 ==========
-The entire GNRC TCP test consists of several test cases. Each test is ran via its own python script
-in the tests directory.
-
-1) 01-conn_lifecycle_as_client.py
-    This test covers TCP connection establishment and teardown with GNRC_TCP acting as tcp client.
-
-2) 02-conn_lifecycle_as_server.py
-    This test covers TCP connection establishment and teardown with GNRC_TCP acting as tcp server.
-
-3) 03-send_data.py
-    This test covers sending of a byte stream from GNRC_TCP to the host system.
-    The amount of data to send is large enough to force GNRC_TCP to split the given stream into
-    multiple packets.
-
-4) 04-receive_data.py
-    This test covers receiving of a byte stream from the host system. The received data is
-    causing window opening and closing as well as data transmission over multiple packets.
-
-5) 05-garbage-pkts.py
-    This test mostly is a regression test for issues that were found through fuzzing. It uses
-    `scapy` to interact with the node.
-
-6) 06-receive_data_closed_conn.py
-    This test covers accessing received data after receiving a FIN packet. If the connection was closed
-    by the peer, a call to gnrc_tcp_recv must return directly with all currently received data
-    or zero if there is no data. The function must return immediately despite any given timeout.
-
-7) 07-endpoint_construction.py
-    This test ensures the correctness of the endpoint construction.
-
-8) 08-return_codes.py
-    This test tries to trigger all documented return codes from GNRC_TCPs functions.
-
-9) 09-listen_accept_cycle.py
-    This test verifies that connection establishment via listen and accept can be repeated multiple
-    times.
-
+The GNRC TCP test test all phases of a tcp connections lifecycle as a server or a client
+as well as TCP behavior on incoming malformed packets.
 
 Setup
 ==========
